requirements.txt
faster-whisper==1.0.3
webrtcvad==2.0.10
sounddevice==0.4.6
numpy==1.26.4
edge-tts==6.1.12
python-dotenv==1.0.1
openai==1.51.0
pydub==0.25.1
